SEO-351 Refactored RobotsTxt extension
* The `wikia-robots-txt.php` logic is now moved to `WikiaRobots`.
* `RobotsTxt` is now renamed to `RobotsTxtBuilder` and reads no context from outside
* `PathBuilder` was extracted with responsibility to build paths for `RobotsTxtBuilder`
* `PathBuilder` reads `$wgContLang` and `$wgArticlePath` to build the paths
* `WikiaRobots` reads all sorts of wg vars to control what's blocked and what's allowed
* Unit tests were added to all three components
* Ability to substitute the `robots.txt` file for a specific wiki was removed
* Experimental method `WikiaRobots::getMetaRobotsPolicy` added, ready for SEO-113

This refactoring is a step towards having the `robots.txt` and `<meta name="robots">` controlled through one class: `WikiaRobots`.

The class builds its state in its constructor, computing the blocked namespaces, blocked params and allowed special pages.

It then uses its state to configure a `RobotsBuilder` instance used to generate the `robots.txt` file, but also in the `getMetaRobotsPolicy` method that can be used to query for the expected `<meta>` tag. This means there's one place to configure the behavior robots should follow on Wikia pages.