SEO-376 Expose HTML title object through getWikiVariables call
In order for Mercury-based apps (CMP first?) to build the HTML titles
consistent with the other pages on Wikia, we're exporting an object
called `htmlTitle` in `MercuryApiController::getWikiVariables`.

The format of the object is:

    "htmlTitle": {
        "separator": " - ",
        "parts": ["Xyz Wiki", "Wikia"]
    }

Mercury-based apps need to read the parts, prepend any additional ones
(required unless it's the main page of the wiki) and then construct the
string by joining with the given separator.

The result HTML titles will be like:

* Kermit the Frog - Muppet Wiki - Wikia
* Quick question - Discussions - Fallout Wiki - Wikia
* Future Films - Wookiepedia - Wikia

Titles structured like that are recommended by SEO team. They serve a
number of purposes:

* Cute titles for search results
* Increase brand awareness (always seeing "Wikia" for our pages)
* Page titles are generally more unique this way
* Page title exhibits the relationship between the specific page, the
  area, the wiki and Wikia, for example "Quick question" is a part of
  "Discussions" which is a part of "Fallout Wiki" which is a part of
  "Wikia"