SEO-218 Update Not_a_valid_Wikia page
Building a custom handler for the article.

We'll enable NotAValidWikia extension on community and the code will
take it from there.

The content of the page remains a message. I'm using another label
though to allow easy transition between old and new. The old version
uses 'not_a_valid_Wikia' message, the new uses 'not-a-valid-wikia'
message (dashes and all lowercase).

There's no need for translations as of now, because you're always taken
to the English community anyway (we may want to support different
languages based on a guess from the domain name or user browser setting).

The ?from param is used to construct the query for the search page.
We're removing the last part from the domain plus optionally the second
last part if it's one of org, com, co, and similar ones. We're dropping
the leading "www" too.

The page will need to be removed from community, so the correct (404)
HTTP status is returned.