SUS-573 - Gatherer - fix URLs
This is a good place to start to see, whether the extension could work.

Remaining actions:

* log all HTTP requests sent by this extension
* log its usage to see how often the page is used

Note:

I could not find any code repository for this extension. I do not know
whether it is actively developed.

If there is no simple way to make it working and to know a little bit
about it, we will simply remove it.