SUS-541: fixed counters on D.I.R.T to show more up to date values.
There was wrong Memcached key being used when updating the counters
which didn't update the values properly. Additionally due to a number
formating any operation on numbers grater than 999 would cause counters
to be corrupted: casting "1,094" to an integer produces undesirable results.
Fix for that is to move value representation/conversion to the controller.