SUS-541: Image stats are now cached in separate keys to avoid race condinons.
To allow global cache for image review stats we should avoid race conditions
and do not store state per user (except reviewer stats). To achieve this since
cas() function is not available for us we split the stats into spearate keys
and use incr() decr() to change the counters. This should eliminate the race
condintions and perhaps allow values to be more up to date.