SEO-456 Use /wiki/Local_Sitemap as the local HTML sitemap
Also change the from/to params in local and global params to avoid being
blocking in robots.txt (?from param is blocked to avoid duplicate
content in category pages).

All the customizations of Special:AllPages are now moved to
LocalSitemapSpecialPage class (a subclass of SpecialAllpages), meaning
users get the original MW page in /wiki/Special:AllPages and bots get the
tweaked one in /wiki/Local_Sitemap.