SEO-406 Make video-enabled XML sitemaps generate faster
In order to generate the sitemaps with enhanced information for videos
we query Solr to get the snippet of the article for each of the file
page that contains a video. Turns out most of the video pages don't have
the content, so the code queries Solr for no good reason, Solr responds
with an empty respons and the code falls back to parsing (empty) wiki
text.

In this pull request, we're first querying Solr to get ALL the pages in
NS_FILE that are indexed in Solr and then only expect a snippet from
those that are indexed. This reduces the time to generate the sitemap to
one tenth of the original time for some wikis (callofduty).