SUS-964: Give Phalanx reload requests more time to succeed
In #11186 we're setting a default timeout value of 1 second for all
internal requests made to Phalanx service. However, this is causing all
reload requests (sent upon saving a block) to time out and fail - this
is producing an error message and also causes MediaWiki to send 4 reload
requests for each save, exacerbating the existing performance issues. It
is safe to increase HTTP timeout value for reload requests, since
they're only sent when a block is saved, so they do not affect site
performance.

Relevant tickets:
- https://wikia-inc.atlassian.net/browse/PLATFORM-2385
- https://wikia-inc.atlassian.net/browse/SUS-890
- https://wikia-inc.atlassian.net/browse/SUS-890

@macbre @mixth-sense