SUS-1215 Be more forgiving of older blocks
Older blocks may not have the ps_blocker_hit column set in the phalanx_stats table. Existing code was using the null coalescing operator to check both ps_blocker_hit then ps_blocker_type.  However as this value is sometimes zero as well as null, this did not do the right thing. Also add back in an old block for stats display purposes, but hide it for block creation.