SUS-541 Fixed tests.
This test no longer checks the method working without a ts paramter.  This is because the old behavior found the ts from memcached, created a URL to redirect to, then got ALL the data it would have gotten, then redirected, throwing all that data away, and reloaded it again.  Now if there is no ts param, we get the right value for it, redirect and don't load anything.  This means there's no data for this test to check, but that's what we want.