(PLATFORM-2391) Fail silently on bad titles due to cache
This controller used to be requested on all page loads without a title
parameter. We're getting a large number of requests to this controller
from cached pages from Googlebot which is giving us thousands of the
following error every hour:

    Error: Call to a member function getText() on null in /extensions/
    wikia/PageShare/PageShareController.class.php:24

So, to avoid unnecessary exceptions, let's fail silently on a bad title
parameter and just set the modal title to an empty string for now.