Allow runOnCluster to process wikis that have been deleted
This allows scripts that want to do a full scan of city_list even if the corresponding db for a deleted wiki is gone.

Options --no-db-check and --filter have been added to achieve this.  --no-db-check will cause runOnCluster to continue
even if it cannot 'use' the DB.  The --filter paramter allows passing 'public', 'private' or 'any' which effects the public=X part of the city_list query.

Updated SiteStatus to accept new paramters that runOnCluster provides to inform about missing DBs