SUS-958 | remove extensions/Oversight
We use extensions/3rdparty/Oversight and extensions/wikia/OversightCompanion:

CommonExtensions.php
735-if( !empty( $wgEnableHideRevisionExt ) ) {
736:	include( "$IP/extensions/3rdparty/Oversight/HideRevision.php" );
737-}
--
851-
852:if (!empty($wgEnableOversightCompanionExt)) {
853:	include( "$IP/extensions/wikia/OversightCompanion/OversightCompanion.php" );
854-}

--

Since July 2012 <https://github.com/Wikia/app/blob/dev/extensions/Oversight/OBSOLETE> says:

As of MediaWiki 1.16.0 this extension is obsolete due to core functionality for
revision deletion (oversight). As such, new installations of this extension are
redundant. This extension is not actively maintained, and is only really kept around
as it is still somewhat used on Wikimedia Wikis.

It will be deleted from subversion in the near future.