SUS-1202 | performance improvements for Chat right module
Refresh user list for in-content chat widget only.
The one on the right rail is always loaded fresh - the chat members list is always up to date as it comes with right rail lazy-loaded HTML.

This makes the great majority of 1,4mm+ requests sent daily to /wikia.php?controller=ChatRail&method=GetUsers&format=json not needed.