SUS-2517: Increase uselessly low IPv6 range block limit
Port of https://gerrit.wikimedia.org/r/#/c/11260/
For IPv6, RFC 3177 recommends that a /48 be allocated to every residential
customer, so range blocks larger than /64 (half the number of bits) will
plainly be required. RFC 4692 implies that a very large ISP may be
allocated a /19 if a generous HD-Ratio of 0.8 is used, so we will use that
as our limit. As of 2012, blocking the whole world would require a /4 range.