SUS-1634 | whitelist WikiFactory AJAX requests to pass the read-only check
Before we can implement wgReadOnlyCluster WikiFactory variable we need to make sure that once it's set for c1, it can be unset as well. Currently it's not possible as WebStart checks whether current wiki (i.e. community.wikia.com) is in read-only (and because wgReadOnlyCluster = 'c1' this check returns true).