Feature branch (#13746)
* Current comment section Monobook Anchor

Blog Listing Pages in the Blog: namespace, while in Monobook, throw comment links which use #comments as an anchor for the comment section, instead of the new #WikiaArticleComments anchor. This change is intended to fix that.

* Current comment section Monobook Anchor

Blog Listing Pages in the Blog: namespace, while in Monobook, throw comment links which use #comments as an anchor for the comment section, instead of the new #WikiaArticleComments anchor. This change is intended to fix that.