(PLATFORM-2851) Make Nirvana cache private for logged-in users
This makes the default cache behaviour for Nirvana controllers to use
public cache for anon users and private cache for logged-in users,
rather than relying on varying on cookie. This matches the behaviour
elsewhere in MediaWiki. If the controller doesn't need to be private
for logged-in users, it can be overridden by explicitly setting the
cache as public.