(PLATFORM-3528) Purge HTTP URLs when edits are made over HTTPS
When a user makes an edit, MediaWiki generates a list of URLs to purge
from the CDN. If the user makes the edit over HTTPS, the URLs generated
were all HTTPS. While we separate the cache on Fastly based on the
protocol, when you purge one URL, it purges both HTTP and HTTPS.
However, if you try to purge the HTTPS version of a URL unauthenticated,
it doesn't work, but the HTTP one does. The MediaWiki purge would be
"anonymous" so the HTTPS URL purge wouldn't work at the moment.

As such, we will for now only purge HTTP URLs for article pages.