SUS-5287 | MWHttpRequest::proxySetup - do not use a localhost as a proxy for "local" URLs + use $wgHTTPProxy value only
When MediaWiki is run inside a container it does not have an access to a localhost proxy as nginx is running in a different container. This commit fixes the following error when trying to perform a HTTP request to a "local" URL, i.e. in *.wikia.com domain. Instead of localhost:80 proxy a HTTP border will be used.

"http-curl-error, Failed to connect to localhost port 80: Connection refused"