SUS-4571 | UpdateThumbnailTask is rarely used, remove it
This task has been used for premium video providers. When we were unable to upload a video thumbnail an offline job was scheduled to retry the upload a bit later.

According to Celery this feature is no longer used (well, we only have a few tasks scheduled in the last 7 days).