CrossWikiCore: set the value of promoted_wiki_b field in xwiki Solr core
Special:Search has a threshold of minimum 50 articles that controls whether a given wiki will show up in the cross wiki search results. We're introducing $wgForceWikiIncludeInSearch WikiFactory variable that can override this limit and force a wiki to show up in the results - promoted_wiki_b field in xwiki Solr core will be set to true in such case.