Enable FounderEmails while in CLI mode
The FounderEmails extension is disabled on Community Central in order
to prevent admins from receiving unnecessary emails. The default
value has been changed in WikiFactory. We have to respect that.

Yet, the maintenance script for FounderEmails has to be executed with
SERVER_ID=177 (Community Central) so we detect CLI mode and include
FounderEmails code even when $wgWikiaEnableFounderEmailsExt is false.

Tada! Magic!