SUS-5880 | more resources for mediawiki-tasks
In order to handle ~50% of tasks we need to increase the number of pods that can handle them. Ths PR also tweaks the `cpu` request for php, so that we have more cpu initially (the request is guaranteed) and the autoscaler is a little slower to react (right now it's a bit too eager).