SUS-6166: PHP session handling broken on Kubernetes
Because we were calling fastcgi_finish_request() client received a response before the request fully completed on the backend side. The session data was not yet saved before the next request from the client hit the backend. The second request had the outdated session data.

This caused an issue with notifications that are handled using PHP session