(PLATFORM-3877) Add a noindex robots tag to HTML API pages
In order to allow deeplinking in our mobile apps, api.php URLs are
explicitly allowed in robots.txt. This has lead to our HTML API docs
ending up in the search index. In order to resolve this, while
hopefully not breaking deeplinking, we can use the meta robots tag
to add "noindex,nofollow" to all the HTML API documentation pages
and see if that cleans these pages out of the index.