PLATFORM-4057| use git abbrev for docker image tag instead of git rev-parse
When deploying image to artifactory we check if the image exists
but when somebody deployed given version to dev/sandbox it does and
it has `dev@hash` version embedded into it (since at that time
release tags/branches were not created yet). This causes prod
image to display outdated info (though the hash and code remains correct).