(PLATFORM-4082) Don't check if commenting is enabled when creating blog
This hook was checking if commenting was enabled on a page when creating blog pages
rather than only checking when creating blog comments. This caused the hook to block
creating blog pages by other means such as via the edit API, since commenting could
not be enabled on a blog page that doesn't exist yet.