(PLATFORM-4199) Don't convert external interwiki URLs to HTTPS
This restores the previous behaviour of wfHttpsAllowedForURL which checks
the number of subdomains and only passes for fandom.com, wikia.com, and
wikia.org URLs. We do not support multiple subdomains over HTTPS on any
domain, and this method is used to check before modifying a URL to be
HTTPS-compatible in a number of places. This includes checking before
converting interwiki URLs to be protocol-relative.