LoadBalancer: only wait on connection if it exists already
Previously, waitFor() would only wait on a connection if one was already open.
It would determine whether a connection was opened by checking if the reader
index was already set. As we now set the reader index unconditionally without
creating a connection, we should refactor the check to take this into account.