(PLATFORM-4386) Remove canonical tags from noindexed pages
Canonical tags on noindexed pages provides a mixed signal to Google and
they will prefer the canonical signal over the noindexed signal and spend
crawl budget on the page.

We want to remove the canonical tag on pages that include the noindex tag.
To do this, this change backports the OutputPageAfterGetHeadLinksArray hook
from newer versions of MediaWiki so the same solution may be usable on the
Unified Platform. It's not the cleanest solution, but it is something that
would work on latest MediaWiki without a core change.