PLATFORM | fix undefined constant in ApiResult
The patch for T128209, backported in fac7ff5731cc992e59d5503e1196f45260ae7796,
references a constant that was added to the ApiResult class in 1.24,
thereby causing an error when the relevant code path is executed.

As a fix, use ApiResult::disableSizeCheck and ApiResult::enableSizeCheck
to achieve the same effect as the new constant would.