MAIN-20832 | Fix LyricWiki XML parser
LyricWiki is embedding an XML feed - https://itunes.apple.com/rss/topsongs/limit=10/explicit=true/xml -
on its main page via a template. It seems that LyricWiki is erroring when trying
to parse this feed with "Error: [] operator not supported for strings".

As a fix, modify the XML parser to initialize $tagData with an array to avoid
the problem.