[WikiDetailsService] Cache per-wiki visualization data
WikiDetailsService is heavily used by API calls from recommendations service to
provide wiki metadata, including a representative image. Most of the returned
data is cached, but the image info is always looked up from the DB. As a result,
this code path is responsible for over 25% of sampled queries (via
GlobalFile::loadData method).

As a fix, ensure we also cache the image data. Also fix foreign connection reuse
in GlobalFile.