Improve BatchRefreshLinksForTemplate task performance
BatchRefreshLinksForTemplate is used to refresh (reparse) articles that depend
on another page, such as a file or template. Currently, each such task reparses
up to 500 articles. We have observed that this batch size is too large; it makes
heavy use of CPU resources and may not complete within the maximum execution
time bounds (180 seconds) allocated for web requests on task executors. Even if
it does complete, a long execution time will mean it will starve other tasks
waiting for execution, thus reducing overall throughput of the task queue each
time there is a spike of BatchRefreshLinksForTemplate tasks due to the
invalidation of some popular page.

As a fix, first reduce the batch size of this task tenfold so that it only
parses up to 50 articles per task, rather than 500. This will naturally result
in a higher number of such tasks generated but should result in generally
smoother performance, as smaller, faster tasks should tie up server resources
less and give other tasks a chance to execute. If too many tasks prove to be a
problem, we can move these tasks to a dedicated queue with lower concurrency.

Second, backport deduplication and caching logic from newer versions to reduce
resource wastage, by leveraging the parser cache if an up-to-date cached parser
output may exist, and aborting early if a subsequent edit to the page that
triggered the recursive task has occurred (and spawned its own set of tasks).
Also fetch titles and revisions for each update batch in a single query rather
than separately.