Reinstate cache for community header
DesignSystemCommunityHeaderModel currently fetches navigation wikitext from the
MW message if one was not passed to it, before using NavigationModel to get the
rendered navigation. This causes the navigation to be freshly parsed on each
request, without leveraging the cache.

As a fix, do not pass any wikitext parameter to NavigationModel if one was not
given. This will make NavigationModel fetch the wikitext from the message by
itself, and, crucially, cache the rendered results.