LinksUpdate: Only wait for replication in auto-commit mode
Improve LinksUpdate replication wait logic to only apply to background tasks,
which run in auto-commit mode, to protect replica DBs from being overwhelmed by
a large amount of concurrent links updates executed on the master. If
LinksUpdate is invoked from an user web request, for instance a purge, don't
wait for replication, as it would require prematurely committing the main
transaction round to actually trigger replication. Instead, prefer to maintain
transactional integrity and skip the wait in this scenario.