refreshlinks: Ensure dependent pages are reparsed in the correct context
In fd23d7838d0afdc8db0f00d6430b7b8316d40af3, BatchRefreshLinksForTemplate was
enhanced to optionally reuse a cached parser output when updating links for
dependent pages. In the process, however, the code was mistakenly changed such
that it'd reparse the page in the context of the title (e.g. a template) that
triggered the links update, causing magic words such as {{PAGENAME}} to be
resolved incorrectly.

As a fix, change the code to again use the title of the dependent page when
reparsing it. Also add an additional sanity check for the cached parser output
to avoid reusing a cached value generated from an earlier revision of the
dependent page.