helios: Gracefully handle the absence of client beacon ID
In 048411ea1331c504b58fd5f60e0eb820d309d88f, we have added an extra step of
comparing the beacon ID returned by helios to the beacon ID present in the
client request. However, the beacon ID may not be present in the request if it
was made to MediaWiki by a service rather than an user's browser directly, e.g.
a request by Feeds to its MediaWiki API controller.

This causes a PHP Notice due to the absent beacon ID that displays an HTML error
box on devboxes, causing responses to break. As a fix, only try to validate the
beacon ID if one was present in the request.