Removes second edit event being sent when redirect is left behind
In https://wikia-inc.atlassian.net/browse/DE-4995 we noticed that there are two events being sent for most page renames compared to UCP.

This is due to page rename DataWarehouse code causing two events edit to be sent.
One that's sent always, and another when redirect_id is not empty.
Redirect_id is not empty when an redirect is left behind (selected by default when moving a page)