Make all replica DB hosts available to DB load balancer
Currently, our DB load balancer configuration contains a single entry for
replicas in each cluster, which is the Consul geo query host name for those
replicas. As such, MediaWiki will only attempt to connect to whichever node was
returned first in the A record; if this connection attempt fails due to e.g. a
timeout (which is not uncommon in our k8s setup due to unresolved conntrack
bugs), it will fail the whole request.

Instead, let's resolve the Consul hostname for replica server configurations,
so that all healthy replica servers returned by Consul[1] are made available to
MediaWiki. This should reduce the possibility of triggering a conntrack race
condition, as we can pick a random replica to connect to on each attempt instead
of always connecting to the first entry, and also allow for retries if we do hit
a connection error.

This mostly restores the code removed in
d1995479fcbaf0dccd79a0a1e4b0c208ca65bfa7 and some older patches while porting
the multi-A resolution logic proposed for UCP.