IW-4212 | Allow pages to be moved to the same title as a video
The page rename hook handler registered by our VideoHandlers extension was
running for all page moves, not just for videos, causing an error when someone
tried to move a page to the same title as an existing video. As a solution, make
sure this handler only runs for video file renames.